Add sample validation for AWS Application Load Balancer JWT
So far, JwtSecurityTokenHandler is still throwing exception which complains about Invalid JWE format. It is suspected that the token is possibly not encrypted.

The sample token (testToken) is having 4 dots, thus consists of 5 segments. According to specification, JWE format is consisted of 5 segments. While the usual JWT (which always refers to JWS) is consisted of 3 segments.

More information:
https://github.com/AzureAD/azure-activedirectory-identitymodel-extensions-for-dotnet/blob/7692d12e49a947f68a44cd3abc040d0c241376e6/src/System.IdentityModel.Tokens.Jwt/JwtSecurityTokenHandler.cs#L298
https://medium.facilelogin.com/jwt-jws-and-jwe-for-not-so-dummies-b63310d201a3?gi=887c719eac6c